Finished updating the code and got all tests to pass for the SalesTaxDAO. Next thing to write are the tests for the SalesTaxSeederFacade.

+++ src/main/java/daos/ISalesTaxDAO.java
@@ -5,6 +5,6 @@ import models.database.SalesTaxDBModel;
 public interface ISalesTaxDAO {
     public boolean create(final SalesTaxDBModel model);
     public SalesTaxDBModel readByZipCode(final String zipCode);
-    public int getTotalNumber();
+    public long getTotalNumber();
     public boolean deleteAll();
 }
